qt-bot/requirements.txt
google-generativeai=0.8.3